OT119-21: move to package template and fix compile issue
