+++ src/main/java/com/alkemy/ong/common/mail/template/EmailAddress.java
@@ -1,0 +1,25 @@
+package com.alkemy.ong.common.mail.template;
+
+import com.alkemy.ong.common.mail.IEmailAddress;
+
+public class EmailAddress implements IEmailAddress {
+
+  private final String address;
+  private final String name;
+
+  public EmailAddress(String address, String name) {
+    this.address = address;
+    this.name = name;
+  }
+
+  @Override
+  public String getAddress() {
+    return address;
+  }
+
+  @Override
+  public String getName() {
+    return name;
+  }
+
+}
