+++ src/main/resources/application.properties
@@ -12,5 +12,5 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jackson.serialization.wrap-root-value=false
 #Email configuration
 email.sender.sendgrid.token=foo
-email.sender.from.address=somosmas.ong@outlook.com
-email.sender.from.name=Somos Más
+email.sender.from.address=somosmas-ong@outlook.com
+email.sender.from.name=Somos-Mas
