--- src/test/java/com/alkemy/ong/common/mail/EmailAddressTest.java
@@ -1,25 +1,0 @@
-package com.alkemy.ong.common.mail;
-
-public class EmailAddressTest implements IEmailAddress {
-
-  private String address;
-  private String name;
-
-  @Override
-  public String getAddress() {
-    return address;
-  }
-
-  public void setAddress(String address) {
-    this.address = address;
-  }
-
-  @Override
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-}
