+++ src/test/resources/application.properties
@@ -7,10 +7,5 @@ spring.jpa.properties.hibernate.hbm2ddl.auto=create-drop
 spring.datasource.platform=h2
 #Email test configuration
 email.sender.sendgrid.token=foo
-email.sender.from.address=somosmas.ong@outlook.com
-email.sender.from.name=Somos Más
-email.sender.to.address=foo@bar
-email.sender.to.name=Foo Bar
-email.sender.content=Welcome
-email.sender.contentType=text/html
-email.sender.subject=Hello
+email.sender.from.address=somosmas-ong@outlook.com
+email.sender.from.name=Somos-Mas
